Moved ImageInfo implementation into it's own source file.
