+++ include/vsg/state/ImageInfo.h
@@ -17,6 +17,8 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 namespace vsg
 {
+    extern VSG_DECLSPEC uint32_t computeNumMipMapLevels(const Data* data, const Sampler* sampler);
+
     /// Settings that map to VkDescriptorImageInfo
     class VSG_DECLSPEC ImageInfo
     {
